% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_direction_legend}
\alias{add_direction_legend}
\title{Add direction legend annotation}
\usage{
add_direction_legend(bbox, curvature = 0.3)
}
\arguments{
\item{bbox}{list with xlim and ylim from \code{\link[=get_data_bbox]{get_data_bbox()}}}

\item{curvature}{curvature value used in the main plot}
}
\value{
A list of ggplot2 annotation layers that can be added to a ggplot object
}
\description{
Builds a mini-legend showing the direction convention for curved branches.
The legend displays two curved arrows demonstrating left-to-right and
right-to-left directions.
}
\examples{
bbox <- list(xlim = c(-10, 10), ylim = c(-5, 5))
p <- ggplot2::ggplot() + ggplot2::xlim(bbox$xlim) + ggplot2::ylim(bbox$ylim)
p + add_direction_legend(bbox)
}
