% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_dispersion_legend}
\alias{add_dispersion_legend}
\title{Add dispersion legend annotation}
\usage{
add_dispersion_legend(bbox, curvature = -0.3, text_size = 2.5)
}
\arguments{
\item{bbox}{list with xlim and ylim from \code{get_data_bbox()}}

\item{curvature}{curvature value used in the main plot}

\item{text_size}{text size for the legend labels (default: 2.5)}
}
\value{
A list of ggplot2 annotation layers that can be added to a ggplot object
}
\description{
Builds a mini-legend showing the dispersion convention for curved branches.
The legend displays two curved arrows demonstrating left-to-right and
right-to-left dispersions.
}
\examples{
bbox <- list(xlim = c(-10, 10), ylim = c(-5, 5))
p <- ggplot2::ggplot() + ggplot2::xlim(bbox$xlim) + ggplot2::ylim(bbox$ylim)
p + add_dispersion_legend(bbox)
}
