% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-polygons.R
\name{build_phylo_hpd}
\alias{build_phylo_hpd}
\title{Build HPD polygons for continuous phylogeography}
\usage{
build_phylo_hpd(
  treedata,
  level = 0.8,
  x_prefix = "location1",
  y_prefix = "location2",
  height = "height_median"
)
}
\description{
Build HPD polygons for continuous phylogeography
}
