% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-segments.R
\name{build_phylo_segments}
\alias{build_phylo_segments}
\title{Build phylogeographic branch segments}
\usage{
build_phylo_segments(
  treedata,
  x = "location1",
  y = "location2",
  height = "height_mean",
  digits = 2
)
}
\description{
Build phylogeographic branch segments
}
