% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylo_branches.R
\name{discretize_curves}
\alias{discretize_curves}
\title{Generate discretized curve points from branch data}
\usage{
discretize_curves(data, curvature = 0.25, ncp = 6, include_progress = FALSE)
}
\description{
Generate discretized curve points from branch data
}
