% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylo_branches.R
\name{geom_curve_path}
\alias{geom_curve_path}
\title{Simple curve geom (plotly-compatible alternative to geom_curve)}
\usage{
geom_curve_path(data, mapping = NULL, curvature = 0.25, ncp = 30, ...)
}
\description{
Simple curve geom (plotly-compatible alternative to geom_curve)
}
