% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylo_branches.R
\name{geom_phylo_branches}
\alias{geom_phylo_branches}
\title{Phylogeographic tree branches}
\usage{
geom_phylo_branches(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings, created with ggplot2::aes().}

\item{data}{A data frame or \code{phylo_phylogeo} object containing branches.}

\item{...}{Other arguments passed on to ggplot2::geom_curve.}
}
\description{
Phylogeographic tree branches
}
