% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylo_nodes.R
\name{geom_phylo_nodes}
\alias{geom_phylo_nodes}
\title{Phylogeographic nodes}
\usage{
geom_phylo_nodes(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings, created with ggplot2::aes().}

\item{data}{A data frame or \code{phylo_phylogeo} object containing nodes.}

\item{...}{Other arguments passed on to ggplot2::geom_point.}
}
\description{
Plot all tree nodes as points filled by their age value (same color
scale as HPD polygons).
}
