% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylo_segments.R
\name{geom_phylo_segments}
\alias{geom_phylo_segments}
\title{Phylogeographic tree segments}
\usage{
geom_phylo_segments(mapping = NULL, data = NULL, ...)
}
\description{
Phylogeographic tree segments
}
