% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylo_branches.R
\name{geom_stream_path}
\alias{geom_stream_path}
\title{Stream curve geom with directional effects}
\usage{
geom_stream_path(
  data,
  mapping = NULL,
  curvature = 0.25,
  ncp = 30,
  global_alpha = 0.6,
  ...
)
}
\description{
Stream curve geom with directional effects
}
