% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{ggphylogeo}
\alias{ggphylogeo-package}
\alias{ggphylogeo}
\title{ggphylogeo: ggplot2 tools for continuous phylogeography}
\description{
Visualisation tools for BEAST-style continuous phylogeographic analyses.
}
\author{
\strong{Maintainer}: Your Name \email{your@email}

}
