% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-app.R
\name{ggplotly_phylogeo}
\alias{ggplotly_phylogeo}
\title{Convert ggphylogeo plot to interactive plotly}
\usage{
ggplotly_phylogeo(p, tooltip = "all")
}
\arguments{
\item{p}{A ggplot2 object from autoplot.treedata}

\item{tooltip}{Which aesthetics to include in tooltip (default: all)}
}
\value{
A plotly object
}
\description{
Takes a ggplot2 object created by autoplot.treedata and converts it to
an interactive plotly visualization.
}
\examples{
\dontrun{
library(treeio)
tree <- read.beast("path/to/beast.tree")
p <- autoplot(tree, most_recent_sample = 2019)
ggplotly_phylogeo(p)
}
}
