% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guides_phylogeo}
\alias{guides_phylogeo}
\title{guides for ggphylogeo plots}
\usage{
guides_phylogeo()
}
\value{
A ggplot2 guide object
}
\description{
A ggplot2 guide tailored for phylogeographic plots.
}
\examples{
ggplot2::ggplot() + guides_phylogeo()
}
