% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h3n2_tree}
\alias{h3n2_tree}
\title{Influenza A H3N2 antigenic cartography tree}
\format{
A treedata object with continuous trait annotations including:
\describe{
\item{antigenic1}{First antigenic dimension coordinate}
\item{antigenic2}{Second antigenic dimension coordinate}
\item{height_mean}{Mean node height (time before present)}
\item{height_median}{Median node height}
\item{antigenic1_0.80HPD}{80 percent HPD intervals for dimension 1}
\item{antigenic2_0.80HPD}{80 percent HPD intervals for dimension 2}
}
}
\source{
BEAST continuous trait analysis
}
\usage{
h3n2_tree
}
\description{
A treedata object containing the maximum clade credibility (MCC) tree
from a BEAST continuous trait analysis of Influenza A H3N2 sequences.
This dataset demonstrates using ggphylogeo with non-geographic continuous
traits (antigenic dimensions).
}
\examples{
\dontrun{
data(h3n2_tree)
pgeo <- build_phylogeo(h3n2_tree, lon = "antigenic1", lat = "antigenic2")
}
}
\keyword{datasets}
