% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_phylogeo}
\alias{theme_phylogeo}
\title{Theme for ggphylogeo plots}
\usage{
theme_phylogeo()
}
\value{
A ggplot2 theme object
}
\description{
A ggplot2 theme tailored for phylogeographic plots.
}
\examples{
ggplot2::ggplot() + theme_phylogeo()
}
