% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wnv_tree}
\alias{wnv_tree}
\title{West Nile Virus phylogeographic tree}
\format{
A treedata object with phylogeographic annotations including:
\describe{
\item{location1}{Latitude coordinate}
\item{location2}{Longitude coordinate}
\item{height_mean}{Mean node height (time before present)}
\item{height_median}{Median node height}
\item{location1_0.80HPD}{80 percent HPD intervals for latitude}
\item{location2_0.80HPD}{80 percent HPD intervals for longitude}
}
}
\source{
BEAST phylogeographic analysis
}
\usage{
wnv_tree
}
\description{
A treedata object containing the maximum clade credibility (MCC) tree
from a BEAST continuous phylogeographic analysis of West Nile Virus
sequences in North America.
}
\examples{
\dontrun{
data(wnv_tree)
autoplot(wnv_tree, most_recent_sample = "2007-07-01")
}
}
\keyword{datasets}
