% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yfv_tree}
\alias{yfv_tree}
\title{Yellow Fever Virus phylogeographic tree}
\format{
A treedata object with phylogeographic annotations including:
\describe{
\item{location1}{Latitude coordinate}
\item{location2}{Longitude coordinate}
\item{height_mean}{Mean node height (time before present)}
\item{height_median}{Median node height}
\item{location1_0.80HPD}{80 percent HPD intervals for latitude}
\item{location2_0.80HPD}{80 percent HPD intervals for longitude}
}
}
\source{
BEAST phylogeographic analysis with relaxed random walk model
}
\usage{
yfv_tree
}
\description{
A treedata object containing the maximum clade credibility (MCC) tree
from a BEAST relaxed random walk (RRW) continuous phylogeographic
analysis of Yellow Fever Virus sequences in South America.
}
\examples{
\dontrun{
data(yfv_tree)
autoplot(yfv_tree)
}
}
\keyword{datasets}
