% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-branches.R
\name{build_branches}
\alias{build_branches}
\title{Build phylogeographic branch segments}
\usage{
build_branches(
  treedata,
  lon = "location1",
  lat = "location2",
  height = "height_mean",
  digits = 2,
  debug = getOption("ggphylogeo.debug", FALSE)
)
}
\arguments{
\item{treedata}{treedata object with phylogeographic annotations}

\item{lon}{name of longitude column (default: "location1")}

\item{lat}{name of latitude column (default: "location2")}

\item{height}{column name used for branch heights (default: "height_mean")}

\item{digits}{number of decimal places for rounding coordinates (default: 2)}

\item{debug}{logical debug messages}
}
\description{
Build phylogeographic branch segments
}
