% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_phylo_hpd.R
\name{geom_phylo_hpd}
\alias{geom_phylo_hpd}
\title{Phylogeographic HPD polygons}
\usage{
geom_phylo_hpd(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mappings, created with ggplot2::aes().}

\item{data}{A data frame or \code{phylo_phylogeo} object containing HPD polygons.}

\item{...}{Other arguments passed on to ggplot2::geom_polygon.}
}
\description{
Phylogeographic HPD polygons
}
