% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{ggphylogeo-package}
\alias{ggphylogeo}
\alias{ggphylogeo-package}
\title{ggphylogeo: ggplot2 tools for continuous phylogeography}
\description{
Visualisation tools for BEAST-style continuous phylogeographic analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Joon-Klaps/ggphylogeo}
  \item Report bugs at \url{https://github.com/Joon-Klaps/ggphylogeo/issues}
}

}
\author{
\strong{Maintainer}: Joon Klaps \email{joon.klaps@example.com}

}
\keyword{internal}
