% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h3n2_tree}
\alias{h3n2_tree}
\title{Influenza A H3N2 continuous trait tree}
\format{
A treedata object with phylogeographic annotations including:
\describe{
\item{antigentic1}{trait one}
\item{antigentic2}{trait two}
\item{height_mean}{Mean node height (time before present)}
\item{height_median}{Median node height}
\item{antigentic1_0.80HPD}{80 percent HPD intervals for trait one}
\item{antigentic2_0.80HPD}{80 percent HPD intervals for trait two}
}
}
\source{
BEAST continuous trait analysis
}
\usage{
h3n2_tree
}
\description{
A treedata object containing the maximum clade credibility (MCC) tree
from a BEAST continuous trait analysis of Influenza A H3N2 sequences.
}
\examples{
\dontrun{
data(h3n2_tree)
autoplot(h3n2_tree, most_recent_sample = "2015-12-31", lon = "antigentic1", lat = "antigentic2")
}
}
\keyword{datasets}
